refactor: split server up into middleware and server
This also adds the new `expo-atlas/middleware` export, to keep server-related code separated from the webui within the plugin. Might remove later when webui imports actual paths within the plugin.
