+++ src/utils/global.ts
@@ -1,0 +1,6 @@
+import { StatsSource } from '../data/types';
+
+declare global {
+  /** The globally initialized data source for Atlas */
+  var EXPO_ATLAS_SOURCE: StatsSource; // eslint-disable-line no-var
+}
